// src/api/generated/model/driverRevenueOption.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */
import type { DriverRevenueOptionEntryCategory } from './driverRevenueOptionEntryCategory';

export interface DriverRevenueOption {
  entryCategory?: DriverRevenueOptionEntryCategory;
  flatRateTypeId?: number;
  label?: string;
  defaultPrice?: number;
}